moved KeenUiSelect to Kselect and removed unused code
